added delete button

+++ blog/views.py
@@ -42,8 +42,11 @@ def post_edit(request, pk):
 
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()
-    return redirect('post_list')
+    if request.method == 'POST':
+        post.delete()
+        return redirect('post_list')
+    # post.delete()
+    # return redirect('post_list')
 
 # def post_remove(request, slug=None, id=None):
 #     if slug:
